requirements.txt
fastapi==0.115.0
uvicorn==0.37.0
requests>=2.32.5
python-dotenv==1.0.1
groq==0.32.0
langchain==0.3.27
langchain-community==0.3.31
langchain-groq==0.3.8
langchain-huggingface==0.3.1
faiss-cpu==1.8.0.post1
transformers==4.44.2
torch==2.4.1
sentence-transformers==3.2.0
